+++ Sources/DexcareiOSSDK/Common/Collection+Extension.swift
@@ -3,7 +3,7 @@
 // ExpressCareCore
 //
 // Created by Reuben Lee on 2018-01-18.
-// Copyright © 2018 Providence. All rights reserved.
+// Copyright © 2018 DexCare. All rights reserved.
 //
 
 import Foundation