+++ Sources/DexcareiOSSDK/DataModels/Common/Gender.swift
@@ -1,4 +1,4 @@
-// Copyright © 2020 Providence. All rights reserved.
+// Copyright © 2020 DexCare. All rights reserved.
 
 /// A generic representation of a Gender
 @frozen