+++ Sources/DexcareiOSSDK/DataModels/Configs/BlisseyConfigs.swift
@@ -1,0 +1,13 @@
+//
+//  BlisseyConfigs.swift
+//  DexcareiOSSDK
+//
+//  Created by Daniel Johns on 2024-02-13.
+//  Copyright © 2024 DexCare. All rights reserved.
+//
+
+import Foundation
+
+struct BlisseyConfigs: Decodable {
+    let minimumWaitTimeForWaitOffline: Int
+}